Szybka refaktoryzacja
